feat: 좋아요를 홀수 번 누르면 DB 에 새로운 좋아요 추가가 아니라 이미 존재하는 좋아요를 삭제되지 않음 상태로 변경하도록 변경
- 한 포스트에서 좋아요가 처음 눌리는 것이 아니지만 이전까지 짝수 횟수만큼 눌려서 삭제된 상태일 때 새로운 좋아요 row 를 추가하지 않고 deleted_at 만 null 로 업데이트하도록 저장된 좋아요 row 의 아이디를 그대로 사용하도록 변경

+++ src/main/java/com/example/likelionfinalproject/service/LikeService.java
@@ -44,9 +44,12 @@ public class LikeService {
 
         Optional<Like> duplicateLike = likeRepository.findByPostIdAndUserId(postId, user.getId());
 
-        if (duplicateLike.isPresent()) {
-            likeRepository.delete(duplicateLike.get());
-            return "좋아요를 해제했습니다.";
+        if (duplicateLike.isPresent() ) {
+            if (duplicateLike.get().getDeletedAt() == null) {
+                likeRepository.delete(duplicateLike.get());
+                return "좋아요를 해제했습니다.";
+            }
+            like.setId(duplicateLike.get().getId());
         }
         likeRepository.save(like);
         return "좋아요를 눌렀습니다.";
